# docs/swagger.yaml
basePath: /
definitions:
  go_auth_internal_apierror.Body:
    properties:
      error:
        $ref: '#/definitions/go_auth_internal_apierror.Detail'
    type: object
  go_auth_internal_apierror.Detail:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
...
  internal_application.Application:
    properties:
      active:
        type: boolean
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      slug:
        type: string
      uuid:
        type: string
    type: object
  internal_application.CreateRequest:
    properties:
      description:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      slug:
        maxLength: 50
        minLength: 2
        type: string
    required:
    - name
    - slug
    type: object
  internal_application.UpdateRequest:
    properties:
      active:
        type: boolean
      description:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      slug:
        maxLength: 50
        minLength: 2
        type: string
    required:
    - name
    - slug
    type: object
  internal_permission.CreateRequest:
    properties:
      application_id:
        type: integer
      description:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      slug:
        maxLength: 50
        minLength: 2
        type: string
    required:
    - application_id
    - name
    - slug
    type: object
  internal_permission.Permission:
    properties:
      active:
        type: boolean
      application_id:
        type: integer
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      slug:
        type: string
    type: object
  internal_permission.UpdateRequest:
    properties:
      active:
        type: boolean
      description:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      slug:
        maxLength: 50
        minLength: 2
        type: string
    required:
    - name
    - slug
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API de autenticação e autorização em Go com JWT, suportando múltiplas
    aplicações
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Go Auth API
  version: "1.0"
paths:
...
  /api/admin/applications:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all applications in the system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_application.Application'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/go_auth_internal_apierror.Body'
      security:
      - Bearer: []
      summary: Get all applications
      tags:
      - Admin - Applications
    post:
      consumes:
      - application/json
      description: Create a new application in the system
      parameters:
      - description: Application data
        in: body
        name: application
        required: true
        schema:
          $ref: '#/definitions/internal_application.CreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_application.Application'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/go_auth_internal_apierror.Body'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/go_auth_internal_apierror.Body'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/go_auth_internal_apierror.Body'
      security:
      - Bearer: []
      summary: Create a new application
      tags:
      - Admin - Applications
  /api/admin/applications/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an application by ID
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/go_auth_internal_apierror.Body'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/go_auth_internal_apierror.Body'
      security:
      - Bearer: []
      summary: Delete an application
      tags:
      - Admin - Applications
    get:
      consumes:
      - application/json
      description: Retrieve a specific application by its ID
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_application.Application'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/go_auth_internal_apierror.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/go_auth_internal_apierror.Body'
      security:
      - Bearer: []
      summary: Get application by ID
      tags:
      - Admin - Applications
    put:
      consumes:
      - application/json
      description: Update an existing application's fields
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: integer
      - description: Application data
        in: body
        name: application
        required: true
        schema:
          $ref: '#/definitions/internal_application.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/go_auth_internal_apierror.Body'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/go_auth_internal_apierror.Body'
      security:
      - Bearer: []
      summary: Update an application
      tags:
      - Admin - Applications
...
  /api/admin/permissions:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all permissions, optionally filtered by application_id
      parameters:
      - description: Filter by application ID
        in: query
        name: application_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_permission.Permission'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/go_auth_internal_apierror.Body'
      security:
      - Bearer: []
      summary: Get all permissions
      tags:
      - Admin - Permissions
    post:
      consumes:
      - application/json
      description: Create a new permission associated with an application
      parameters:
      - description: Permission data
        in: body
        name: permission
        required: true
        schema:
          $ref: '#/definitions/internal_permission.CreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_permission.Permission'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/go_auth_internal_apierror.Body'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/go_auth_internal_apierror.Body'
      security:
      - Bearer: []
      summary: Create a new permission
      tags:
      - Admin - Permissions
  /api/admin/permissions/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a permission by ID
      parameters:
      - description: Permission ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/go_auth_internal_apierror.Body'
      security:
      - Bearer: []
      summary: Delete a permission
      tags:
      - Admin - Permissions
    get:
      consumes:
      - application/json
      description: Retrieve a specific permission by its ID
      parameters:
      - description: Permission ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_permission.Permission'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/go_auth_internal_apierror.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/go_auth_internal_apierror.Body'
      security:
      - Bearer: []
      summary: Get permission by ID
      tags:
      - Admin - Permissions
    put:
      consumes:
      - application/json
      description: Update an existing permission's fields
      parameters:
      - description: Permission ID
        in: path
        name: id
        required: true
        type: integer
      - description: Permission data
        in: body
        name: permission
        required: true
        schema:
          $ref: '#/definitions/internal_permission.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/go_auth_internal_apierror.Body'
      security:
      - Bearer: []
      summary: Update a permission
      tags:
      - Admin - Permissions
...
securityDefinitions:
  Bearer:
    description: Digite "Bearer " seguido do token JWT
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"